just add

+++ ft_printf/case_conversion/case_conversion.h
@@ -5,9 +5,8 @@ int	case_s(char *s, t_conversion_setting *conversion_setting);
 int	case_di(int n, t_conversion_setting *conversion_setting);
 int	case_c(char c, t_conversion_setting *conversion_setting);
 int	case_p(void *p, t_conversion_setting *conversion_setting);
-int	case_u(long long int n, t_conversion_setting *conversion_setting);;
-int	case_xX(long long int n, t_conversion_setting *conversion_setting);
+int	case_u(unsigned int n, t_conversion_setting *conversion_setting);;
+int	case_xX(unsigned long long int n, t_conversion_setting *conversion_setting);
 int	print_n_c(char c, int len);
-
 
 #endif
